[refactor] Moved "isString()" to its own TS file
